+++ Engine_SOURCE/jkPaintShader.cpp
@@ -7,8 +7,8 @@ namespace jk::graphics
 	{
 		mTarget->BindUnorderedAccessViews(0);
 
-		mGroupX = mTarget->GetWidth() / mThreadGroupCountX + 1;
-		mGroupY = mTarget->GetHeight() / mThreadGroupCountY + 1;
+		mGroupX = static_cast<UINT>(mTarget->GetWidth()) / mThreadGroupCountX + 1;
+		mGroupY = static_cast<UINT>(mTarget->GetHeight()) / mThreadGroupCountY + 1;
 		mGroupZ = 1;
 	}
 	void PaintShader::Clear()
